fix issue with seeds overflowing
now I remember why I made those longs

+++ randomizer_bruteforce/all_default/one_hub_F6/Run.java
@@ -33,8 +33,8 @@ class Run extends RunnableThread {
     private static int THREAD_NUM = 8;
     private static int PER_LOOP = 100000;
     private static int PER_THREAD = (PER_LOOP / THREAD_NUM);
-    private static int current_seed = 0;
-    private static int max_seed = 0x7FFFFFFF;
+    private static long current_seed = 0;
+    private static long max_seed = 0x7FFFFFFF;
 
     public static void main(String[] args) {
         // So Ctrl-C gives some output
